+++ es/resource_elasticsearch_opendistro_ism_policy_test.go
@@ -128,6 +128,10 @@ resource "elasticsearch_opendistro_ism_policy" "test_policy" {
 		"policy": {
 		  "description": "ingesting logs",
 		  "default_state": "ingest",
+      "ism_template": {
+        "index_patterns": ["foo-*"],
+        "priority": 0
+			},
 		  "error_notification": {
         "destination": {
           "slack": {