+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 ### Fixed
+- [opendistro ism policy] Fix perpetual diff in ism_template if not specified
 
 
 ## [1.5.6] - 2020-05-10